Set all config in PHP
Before: Most of the default config was set in the JS. This meant that if
we wanted to e.g. customise the button style of the cookie control
panel, the only way to do so was by editing the JS of this plugin.

Now: All the config is set in the PHP. This will allow us to (in an
upcoming change) add filters to the config, which means other plugins or
themes could hook into this plugin to change its default config.