Output GA4 script tag if GA4 ID is set
Because Google recommend to add an async attribute to this tag, we can't
use wp_enqueue_script to add it, so instead we're hooking into wp_head
and echoing it out.

We'll also define the `gtag()` function at this point as well, as per
the instructions here:
https://support.google.com/analytics/answer/10271001?hl=en&ref_topic=11192706#analyticsjs-enable-basic&zippy=%2Cin-this-article.

This will enable us to set default consent levels even when a user
hasn't explicitly opted in or out of analytics.