Run basic gtag config as soon as page loads
Before: we were waiting for the DOM to be ready before doing some of the
GA4 setup. In fact, this is not required as data can be pushed to
window.dataLayer and GA4 will just deal with that data once the gtag
script has loaded

Now: we do the basic setup straight away, which allows us to simplify
the JS setup.