Load GTM in the head, rather than if analytics cookies accepted
Before: if GTM was configured, we only loaded it if the user had opted
into analytics cookies. This was fine when GTM would only be loading
analytics scripts, but now we're going to allow users to load marketing
scripts via GTM as well, this won't work.

Now: we load the GTM script in the head if a GTM id is provided. We'll
then just use the opt in/out events to control the GTM consent state.