Add `wp-postpass_*` to necessary cookies array
The `necessaryCookies` array is a list of cookies that should not be
removed by Civic Cookie Control when clearing cookies.

Most cookies WordPress sets get marked as `httponly`, and if so then
Civic Cookie Control does not remove them. However, `wp-postpass_*`
cookies (used to allow access to WordPress password-protected page) do
not get marked as such, which means they get deleted on each page load.
This means users are prompted repeatedly for the password to
password-protected pages, when they should only be asked once. Adding
the `wp-postpass_*` term to the `necessaryCookies` array prevents that
unwanted behaviour.