Add subresource integrity to the Civic script
Previously this plugin used the latest stable version
of the Civic, which meant that we could not provide a
subresource integrity attribute with the 'script' tag.
This means that any site using this plugin would be
vulnerable to CORS style attacks.

In this commit we add the computed SRI value to the
script when it is loaded to avoid CRSF attacks.

See:
    https://www.civicuk.com/cookie-control/documentation
    https://developer.mozilla.org/en-US/docs/Web/Security/Subresource_Integrity